+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # CHANGELOG
 
+## v2.4
+
+* Fix bug in v2.3 that double block still works when the player is on ground.
+* Switching to ladder or hardest block now has a new feature: after some seconds of block-in (default 15s, configurable at `blockin_post_time`), the switch is directly to hardest block, without preferring ladder.
+
 ## v2.3
 
 * Disable block clutch if has jump boost.